Initialise physical output volume settings
Add entries in the model to hard-code some mixer elements to specific
values. Use this to set the headphone outputs to SW-controlled on full
volume so they are only in effect controlled by the potentiometers on
the device.

Other outputs are set to "HW", which means they are controlled by the
"Monitor" physical control.